Revert "feat(payment): STRIPE-476 Stripe Google Pay 3DS added"
