+++ packages/google-pay-integration/src/factories/payment/create-google-pay-stripe-payment-strategy.ts
@@ -1,18 +1,15 @@
 import { createFormPoster } from '@bigcommerce/form-poster';
 import { createRequestSender } from '@bigcommerce/request-sender';
-import { getScriptLoader } from '@bigcommerce/script-loader';
 
 import {
     PaymentStrategyFactory,
     toResolvableModule,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
 
-import createGooglePayScriptLoader from '../factories/create-google-pay-script-loader';
-import GooglePayPaymentProcessor from '../google-pay-payment-processor';
-import GooglePayPaymentStrategy from '../google-pay-payment-strategy';
-
-import GooglePayStripeGateway from './google-pay-stripe-gateway';
-import StripeUPEScriptLoader from './stripe-upe-script-loader';
+import GooglePayStripeGateway from '../../gateways/google-pay-stripe-gateway';
+import GooglePayPaymentProcessor from '../../google-pay-payment-processor';
+import GooglePayPaymentStrategy from '../../google-pay-payment-strategy';
+import createGooglePayScriptLoader from '../create-google-pay-script-loader';
 
 const createGooglePayStripePaymentStrategy: PaymentStrategyFactory<GooglePayPaymentStrategy> = (
     paymentIntegrationService,
@@ -21,10 +18,7 @@ const createGooglePayStripePaymentStrategy: PaymentStrategyFactory<GooglePayPaym
         paymentIntegrationService,
         new GooglePayPaymentProcessor(
             createGooglePayScriptLoader(),
-            new GooglePayStripeGateway(
-                paymentIntegrationService,
-                new StripeUPEScriptLoader(getScriptLoader()),
-            ),
+            new GooglePayStripeGateway(paymentIntegrationService),
             createRequestSender(),
             createFormPoster(),
         ),
