--- packages/google-pay-integration/src/google-pay-stripe/stripe.mock.ts
@@ -1,8 +1,0 @@
-import { StripeUPEClient } from './types';
-
-export function getStripeUPEJsMock(): StripeUPEClient {
-    return {
-        confirmCardPayment: jest.fn(),
-        retrievePaymentIntent: jest.fn(),
-    };
-}
