--- packages/google-pay-integration/src/guards/is-google-pay-stripe-request-error.ts
@@ -1,12 +1,0 @@
-interface AdditionalActionError {
-    three_ds_result: { token: string };
-    errors: Array<{ code: string }>;
-}
-
-const isGooglePayStripeRequestError = (body: unknown): body is AdditionalActionError => {
-    return (
-        typeof body === 'object' && body !== null && 'three_ds_result' in body && 'errors' in body
-    );
-};
-
-export default isGooglePayStripeRequestError;
